pyproject.toml
[tool.poetry]
name = "fantasy_football"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
pandas = "^1.4.3"
scipy = "^1.9.0"
numpy = "^1.23.1"
requests = "^2.28.1"
psycopg = "^3.0.16"
psycopg2-binary = "^2.9.3"
SQLAlchemy = "^1.4.40"
PuLP = "^2.6.0"
wordpress-api = "^1.2.9"
matplotlib = "^3.5.3"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
